v2_legacy/requirements.txt
streamlit
biopython
openai
sentence-transformers
chromadb
google-generativeai

requests
beautifulsoup4
matplotlib
python-dotenv

# PDF processing (for local Read functionality)
pymupdf
layoutparser
pdf2image
opencv-python
torch
torchvision
python-pptx==1.0.2
PyMuPDF==1.26.7
